+++ include/websocket.h
@@ -98,6 +98,6 @@ void ws_write_frame(socket_struct *ns, SockList *sl);
  * @param out_len  Size of @p out (must be at least 29 bytes).
  * @return         -1 in case of error, or out size, including terminating byte.
  */
-ssize_t ws_compute_accept_key(const char *key, char *out, size_t out_len);
+int ws_compute_accept_key(const char *key, char *out, size_t out_len);
 
 #endif /* WEBSOCKET_H */
